Done unit-tests for lab 2

+++ 2/strophoid.h
@@ -6,21 +6,21 @@ namespace strophoid {
 
 	public:
 
-		Strophoid(const long double A = 1.0);
+		Strophoid(const double A = 1.0);
 
-		void setA(const long double A = 1.0);
+		void setA(const double A = 1.0);
 		
-		long double getA() const;
-		long double getY(const long double X = 0) const;
-		long double getRadius(long double angle = 0) const;
-		long double getRadiusOfCurvature() const;
-		long double getLoopArea() const;
-		long double getLoopVolume() const;
-		long double getMiddleBranchesArea() const;
+		double getA() const;
+		double getY(const double X = 0) const;
+		double getRadius(double angle = 0) const;
+		double getRadiusOfCurvature() const;
+		double getLoopArea() const;
+		double getLoopVolume() const;
+		double getMiddleBranchesArea() const;
 
 	private:
 
-		long double A;
+		double A;
 
 	};
 }
